Moved the ngram loaders to data_input.
